Exceptions

+++ src/main/java/ru/isaykin/application/services/TruckService.java
@@ -3,6 +3,7 @@ package ru.isaykin.application.services;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
+import ru.isaykin.application.exceptions.NoTruckException;
 import ru.isaykin.application.model.Measure;
 import ru.isaykin.application.model.Truck;
 import ru.isaykin.application.repositories.MeasureRepository;
@@ -38,10 +39,16 @@ public class TruckService {
     }
 
     public Truck getTruck(Long id) {
-        return truckRepository.getById(id);
+        Truck result = truckRepository.getById(id);
+        if (result == null) throw new NoTruckException("There is no Truck in database with such id");
+
+
+        return result;
     }
 
     public void deleteById(Long id) {
+        if (truckRepository.getById(id) == null)
+            throw new NoTruckException("There is no Truck in database with such id");
         List<Measure> measureList = measureRepository.getAll();
         for (Measure measure : measureList) {
             if (measure.getTruckId().equals(id)) measureRepository.deleteById(measure.getId());
@@ -52,6 +59,7 @@ public class TruckService {
     public void updateById(Long id, Truck updatedTruck) {
         Truck truckToUpdate = truckRepository.getById(id);
         if (truckToUpdate == null) {
+            throw new NoTruckException("There is no Truck in database with such id");
         } else {
             updatedTruck.setId(id);
             truckRepository.save(updatedTruck);
